More Refactoring

+++ Blockchain/NSData+Hex.m
@@ -22,7 +22,7 @@
     NSMutableString     *hexString  = [NSMutableString stringWithCapacity:(dataLength * 2)];
     
     for (int i = 0; i < dataLength; ++i)
-        [hexString appendString:[NSString stringWithFormat:@"%02x", (unsigned long)dataBuffer[i]]];
+        [hexString appendString:[NSString stringWithFormat:@"%02x", (unsigned int)dataBuffer[i]]];
     
     return [NSString stringWithString:hexString];
 }
